+++ internal/domain/ports/outbound/repository/profile_repository.go
@@ -7,5 +7,5 @@ import (
 
 type ProfileRepository interface {
 	CreateProfile(ctx context.Context, profile *entities.Profile) (entities.Profile, error)
-	DeleteProfile(ctx context.Context, profileId int32) error
+	DeleteProfile(ctx context.Context, profileId int64) error
 }
